Rename directory from `base_models` to `detectors`
