+++ src/otx/algo/detection/detectors/__init__.py
