+++ src/otx/algo/detection/yolox.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any
 
 from otx.algo.common.losses import CrossEntropyLoss, L1Loss
 from otx.algo.detection.backbones import CSPDarknet
-from otx.algo.detection.base_models import SingleStageDetector
+from otx.algo.detection.detectors import SingleStageDetector
 from otx.algo.detection.heads import YOLOXHead
 from otx.algo.detection.losses import IoULoss
 from otx.algo.detection.necks import YOLOXPAFPN
